Issue #1758. Separate tracker service and calculateStatusEventsSums
